Updated Interval
Updated the type of interval read from the Config.json file from integer to double.

+++ Bot/Converters/TimeSpanConverter.cs
@@ -4,7 +4,7 @@
 	{
 		public override TimeSpan Read(ref Utf8JsonReader reader, Type typeToConvert, JsonSerializerOptions options)
 		{
-			return TimeSpan.FromSeconds(reader.GetInt32());
+			return TimeSpan.FromSeconds(reader.GetDouble());
 		}
 
 		public override void Write(Utf8JsonWriter writer, TimeSpan value, JsonSerializerOptions options)
